requirements.txt
apache-airflow
hdfs
kafka-python
pyspark
apache-airflow-providers-apache-spark
python-dotenv
requests